cast precision specifiers to int (to eliminate warnings)

+++ src/tcl/tclutil.c
@@ -175,9 +175,10 @@ TclFindElement(interp, list, elementPtr, nextPtr, sizePtr, bracePtr)
 			/* null body */
 		    }
 		    Tcl_ResetResult(interp);
+		    /* sde cast to (int) of precision specifier */
 		    sprintf(interp->result,
 			    "list element in braces followed by \"%.*s\" instead of space",
-			    p2-p, p);
+			    (int) (p2-p), p);
 		    return TCL_ERROR;
 		} else if (openBraces != 0) {
 		    openBraces--;
@@ -232,9 +233,10 @@ TclFindElement(interp, list, elementPtr, nextPtr, sizePtr, bracePtr)
 			/* null body */
 		    }
 		    Tcl_ResetResult(interp);
+		    /* sde cast to (int) of precision specifier */
 		    sprintf(interp->result,
 			    "list element in quotes followed by \"%.*s\" %s",
-			    p2-p, p, "instead of space");
+			    (int) (p2-p), p, "instead of space");
 		    return TCL_ERROR;
 		}
 		break;
